testCreateNewViewInFolder move to ViewTest
